utility: Use 64bit type to hold initial table size in meta.

+++ divine/utility/meta.h
@@ -27,7 +27,7 @@ struct Input {
 };
 
 struct Execution {
-    int initialTable;
+    intptr_t initialTable;
     bool diskFifo;
 
     int threads;
